requirements.txt
Flask
Flask-SocketIO
eventlet
mecab-python3
unidic-lite
